Adding the builder to the codebase
The new 'build.php' file serves as both a source builder and a source viewer. Visiting the page allows you to view the source of both the full source and the minified files as well as each animations individual source files before compiling.

There is also a "Rebuild the Source" button above the nav which will automatically compile all of the individual animation source files, as well as create the minified version. It also creates a new Table of Contents and headers for each animation in the full source file, allowing easy navigation through the code.