Prevent double execution of code on animation end
Prevent double execution of code on animation end

Because of some modern browsers, `animationend``will become the new
standards, thus having both, their vendor event and the w3c event
implemented.
It is best to keep track with a variable if the event has ended as well
to prevent double execution.