Remove event names for prehistoric browsers
Looking through http://caniuse.com/#search=animation I saw that "MSAnimationEnd" was never supported in the first place. IE10 was the first version to support CSS animations and it used the un-prefixed event name. "mozAnimationEnd" and "oanimationend" are for browsers from 2012 and earlier with no more than a hundredth of 1 percent usage.